#3 Moves Fixtures out of src dir
