+++ fixtures/index.ts
@@ -1,2 +1,4 @@
 export * from './axiosResponse'
 export * from './stringCase'
+export * from './posts'
+export * from './axiosError'
