+++ fixtures/stringCase.ts
