+++ jest.config.ts
@@ -12,7 +12,6 @@ module.exports = {
   collectCoverageFrom: [
     'src/**/*.{js,jsx,ts,tsx}',
     '!src/pages/api/**/*.(ts|tsx)',
-    '!src/fixtures/**/*.(ts|tsx)',
     '!src/app/layout.tsx',
     '!src/common/types/**/*.(ts|tsx)',
   ],
