+++ src/app/posts/hooks/__tests__/usePosts.test.ts
@@ -2,8 +2,7 @@ import { INTERNAL_API_ROUTES } from '@/common/constants'
 import mockAxios from 'jest-mock-axios'
 import { act, renderHook, waitFor } from '@testing-library/react'
 import { AxiosErrorFacade } from '@/common/facades'
-import { GET_POSTS_RESPONSE } from '@/fixtures/posts'
-import { AXIOS_ERROR } from '@/fixtures/axiosError'
+import { AXIOS_ERROR, GET_POSTS_RESPONSE } from '../../../../../fixtures'
 import { TestSWRConfig } from '../../../../../jest.components'
 import { usePosts } from '..'
 import { PostsDataFacade } from '../../facades'
