+++ src/common/utils/__tests__/camelCaseKeys.test.ts
@@ -1,4 +1,4 @@
-import { CAMEL_CASED, SNAKE_CASED } from '@/fixtures'
+import { CAMEL_CASED, SNAKE_CASED } from '../../../../fixtures'
 import { camelCaseKeys } from '..'
 
 describe('camelCaseKeys', () => {
