+++ src/common/utils/__tests__/snakeCaseKeys.test.ts
@@ -1,4 +1,4 @@
-import { CAMEL_CASED, SNAKE_CASED } from '@/fixtures'
+import { CAMEL_CASED, SNAKE_CASED } from '../../../../fixtures'
 import { snakeCaseKeys } from '..'
 
 describe('snakeCaseKeys', () => {
